requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.0
requests==2.32.4
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
